Moved ForgeTypeEnums to be constant int32 values. This makes it easier to add some internal values without having to do a lot of casting.
Addedd a couple of internal request types. add_asset_preview actually consists of several calls to the api so I needed some internal requests that are only handled by the forge module.

Stuck on an issue when converting a config to json that is causing couchdb to not like the json. Needs to be resovled.